requirements.txt
dlib==19.18.0
imutils==0.5.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
torch==1.3.1
torchvision==0.4.2